+++ src/components/Navbar.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import Input from "components/Input";
-import { fadeIn } from "components/animations";
+import { fadeIn } from "styled";
 
 const Navbar = () => {
   return (
